@tomturner 2022-03-03
Really sorry but I won't be merging this as it will break alot of project including mine.  The reason is that it get the whole tenant model / object is that fact you can get to other fields in the request object.  IE in my project I store what parts people are allowed to access in the tenant model. I then have access to it everywhere though the request.  If you want to do are suggesting I would create your own middleware inherided for the main middleware and that should do the trick.

Thanks Tom